requirement.txt
streamlit
langchain
langchain-community
langchain-groq
groq
python-docx
pymupdf
python-dotenv
